requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
itsdangerous>=2.1
jinja2>=3.1
passlib[bcrypt]>=1.7
bcrypt>=4.0,<4.1
python-dotenv>=1.0
opencv-contrib-python>=4.9
numpy>=1.26
requests>=2.31
